+++ stream-chat-android-ui-common/src/main/kotlin/io/getstream/chat/android/ui/common/feature/messages/list/MessageListController.kt
@@ -224,6 +224,7 @@ public class MessageListController(
     public val unreadLabelState: MutableStateFlow<UnreadLabel?> = MutableStateFlow(null)
     private val showUnreadButtonState = MutableSharedFlow<Boolean>(extraBufferCapacity = 1)
     private val updateUnreadLabelState = MutableStateFlow(true)
+    private val originalTranslationsStore by lazy { MessageOriginalTranslationsStore.forChannel(cid) }
 
     /**
      * Holds information about the abilities the current user is able to exercise in the given channel.
@@ -470,6 +471,7 @@ public class MessageListController(
                 unreadLabelState,
                 channelState.members,
                 channelState.endOfOlderMessages,
+                originalTranslationsStore.originalTextMessageIds,
             ) { data ->
                 val state = data[0] as MessagesState
                 val reads = data[1] as List<ChannelUserRead>
@@ -484,6 +486,7 @@ public class MessageListController(
                 val unreadLabel = data[10] as UnreadLabel?
                 val members = data[11] as List<Member>
                 val endOfOlderMessages = data[12] as Boolean
+                val messagesInOriginalLanguage = data[13] as Set<String>
 
                 when (state) {
                     is MessagesState.Loading,
@@ -518,6 +521,7 @@ public class MessageListController(
                                 read = reads,
                             ),
                             ownCapabilities = channel.ownCapabilities,
+                            messagesInOriginalLanguage = messagesInOriginalLanguage,
                         ),
                         endOfNewMessagesReached = endOfNewerMessages,
                     )
@@ -760,6 +764,7 @@ public class MessageListController(
                 focusedMessage,
                 members,
                 ownCapabilities,
+                originalTranslationsStore.originalTextMessageIds,
             ) { data ->
                 val messages = data[0] as List<Message>
                 val reads = data[1] as List<ChannelUserRead>
@@ -772,6 +777,7 @@ public class MessageListController(
                 val focusedMessage = data[8] as Message?
                 val members = data[9] as List<Member>
                 val ownCapabilities = data[10] as Set<String>
+                val messagesInOriginalLanguage = data[11] as Set<String>
 
                 _threadListState.value.copy(
                     isLoading = false,
@@ -794,6 +800,7 @@ public class MessageListController(
                         endOfOlderMessages = false,
                         channel = null,
                         ownCapabilities = ownCapabilities,
+                        messagesInOriginalLanguage = messagesInOriginalLanguage,
                     ),
                     parentMessageId = threadId,
                     endOfNewMessagesReached = true,
@@ -857,6 +864,7 @@ public class MessageListController(
         endOfOlderMessages: Boolean,
         channel: Channel?,
         ownCapabilities: Set<String>,
+        messagesInOriginalLanguage: Set<String>,
     ): List<MessageListItemState> {
         val parentMessageId = (_mode.value as? MessageMode.MessageThread)?.parentMessage?.id
         val currentUser = user.value
@@ -950,6 +958,7 @@ public class MessageListController(
                         messageReadBy = messageReadBy,
                         focusState = if (isMessageFocused) MessageFocused else null,
                         ownCapabilities = ownCapabilities,
+                        showOriginalText = messagesInOriginalLanguage.contains(message.id),
                     ),
                 )
             }
@@ -1234,6 +1243,15 @@ public class MessageListController(
             reads = channelState.reads,
             members = channelState.members,
         )
+    }
+
+    /**
+     * Toggles between the translated and the original text of the message (if the message was auto-translated).
+     *
+     * @param messageId The ID of the message for which to toggle the original text.
+     */
+    public fun toggleOriginalText(messageId: String) {
+        originalTranslationsStore.toggleOriginalText(messageId)
     }
 
     /**
@@ -2335,7 +2353,7 @@ public class MessageListController(
      */
     public fun onCleared() {
         // Clear any messages for which the original text was shown
-        MessageOriginalTranslationsStore.forChannel(cid).clear()
+        originalTranslationsStore.clear()
         scope.cancel()
     }