+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/feature/messages/list/adapter/internal/MessageListItemDiffCallback.kt
@@ -78,6 +78,7 @@ internal object MessageListItemDiffCallback : DiffUtil.ItemCallback<MessageListI
             footer = oldItem.showMessageFooter != newItem.showMessageFooter,
             poll = oldMessage.poll != newMessage.poll,
             threadMode = oldItem.isThreadMode != newItem.isThreadMode,
+            showOriginalText = oldItem.showOriginalText != newItem.showOriginalText,
         )
     }