+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/feature/messages/list/adapter/viewholder/impl/GiphyAttachmentViewHolder.kt
@@ -81,6 +81,9 @@ public class GiphyAttachmentViewHolder internal constructor(
                 footnote.setOnThreadClickListener {
                     container.threadClickListener.onThreadClick(data.message)
                 }
+                footnote.setOnTranslatedLabelClickListener {
+                    container.translatedLabelClickListener.onTranslatedLabelClick(data.message)
+                }
                 messageContainer.setOnLongClickListener {
                     container.messageLongClickListener.onMessageLongClick(data.message)
                     true