+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/utils/extensions/MessageListItem.kt
@@ -49,6 +49,7 @@ public fun MessageListItemCommon.toUiMessageListItem(): MessageListItem {
             isThreadMode = isInThread,
             isMessageRead = isMessageRead,
             showMessageFooter = showMessageFooter,
+            showOriginalText = showOriginalText,
         )
         is EmptyThreadPlaceholderItemState -> MessageListItem.ThreadPlaceholderItem
         is UnreadSeparatorItemState -> MessageListItem.UnreadSeparatorItem(unreadCount = unreadCount)